Update site metadata and enforce code consistency
Enhanced the site's metadata in the layout component with a personalized title and description, reflecting the shift towards a more personal portfolio and blog for Thibault Guilhem. Adjusted the language attribute to French, aligning with the target audience. Additionally, introduced '.txt' files to .gitignore to prevent clutter in the repository, and created a 'notes.txt' file, possibly for project ideas or planning purposes. Code styling in 'layout.js' has been made consistent with the addition of semicolons, promoting better readability and maintenance.